requirements.txt
amqp==2.5.2
asgiref==3.2.5
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
cryptography==2.9.2
cssselect2==0.3.0
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.4
django-redis==4.11.0
djangorestframework==3.11.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
kombu==4.6.8
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
Pyphen==0.9.5
pytz==2019.3
redis==3.4.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
Twisted==20.3.0
txaio==20.4.1
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
zipp==3.1.0
zope.interface==5.1.0
